requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
lightgbm==4.1.0
xgboost==2.0.3
joblib==1.3.2
fastapi==0.104.1
uvicorn==0.24.0
shap==0.44.0
openai==1.3.7
rapidfuzz==3.6.1
azure-ai-formrecognizer==3.3.0
evidently==0.4.0
mlflow==2.8.1